refactor: address review todos
- Replace the PR-invented "marker"/"marking" nomenclature with the
  terms the project already uses: "generated class" in docs/READMEs,
  "postfix class" in code (matching the postfix variable)
- Move attribute-selector rewrite tests into rewrite.attributes.test.ts
  and extract selectors via postcss instead of regex
- Rename rewriteHbs.test.ts to rewriteHbs.attributes.test.ts
- Compare Sets directly in utils.js tests instead of spreading to arrays
- Collapse elementHasScopedAttribute to a one-liner: the attributes set
  comes from CSS parsing, so @args/...attributes can never be in it
- Document why splitting [class="foo bar"] is manual: the selector
  parser exposes attribute values only as opaque strings
